# OPENAPI_SPEC.yaml
openapi: 3.0.3
info:
  title: RiskModels API
  version: 3.0.0-agent
  description: >
    Precision equity risk intelligence platform with AI agent integration.

    Institutional-grade equity risk analysis for AI agents and quantitative applications.
    Provides daily factor hedge ratios, explained-risk decompositions, and portfolio-level
    risk attribution for ~3,000 US equities.


    **New in v3.0.0-agent:**

    - OAuth2 client credentials flow for machine-to-machine auth

    - Plaid Investments integration for live portfolio sync

    - MCP (Model Context Protocol) server for AI agent tools

    - Compliance manifests for AI marketplace integration

    - Per-key rate limiting and scoped access control


    **Data**: Universe `uni_mc_3000` (top ~3,000 US stocks by market cap), date range
    2006-01-04 to present, updated daily end-of-day. Backend: Zarr v2 on Google Cloud Storage,
    computed by the ERM3 regression system (Huber/Ridge). Supabase tables `erm3_betas` and
    `erm3_rankings` provide factor betas and ticker rankings for direct DB and API use.


    **Authentication Methods:**

    1. Session cookies (web app)

    2. API Key/Secret via Bearer token (external integrations)

    3. OAuth2 client credentials (AI agents, npm package)


    **Security Features:**

    - GCP KMS envelope encryption for portfolio data

    - Cryptographic shredding on key deletion

    - HMAC-SHA256 webhook verification

    - Row-level security (RLS) via Supabase


...

...

...

...

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: rm_agent_{environment}_{random}_{checksum}
      description: >
        API key in format `rm_agent_{live|test}_{random}_{checksum}` or `rm_user_{random}_{checksum}`.
        Obtain via `POST /api/auth/provision` or from riskmodels.app/settings → API Keys.
        Include as: `Authorization: Bearer rm_agent_live_...`

    OAuth2ClientCredentials:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: https://riskmodels.app/api/auth/token
          scopes:
            ticker-returns: Access ticker returns and historical data
            risk-decomposition: Access L3 risk decomposition
            batch-analysis: Perform portfolio batch analysis
            factor-correlation: Correlate stocks with macro factors (VIX, Bitcoin, Gold, etc.)
            chat-risk-analyst: Use AI risk analyst
            plaid:holdings: Access Plaid-synced portfolio holdings
            "*": Full API access
...
  schemas:
    AgentMeta:
      type: object
      description: Per-request billing and telemetry metadata appended to metered responses.
      properties:
        cost_usd:
          type: number
          format: float
          description: Cost deducted from prepaid balance for this request. 0 if served from cache.
          example: 0.005
        latency_ms:
          type: integer
          description: Server processing time in milliseconds.
          example: 145
        request_id:
          type: string
          description: Unique request identifier for support lookups.
          example: req_abc123xyz
        confidence:
          type: object
          properties:
            overall:
              type: number
              format: float
              minimum: 0
              maximum: 1
              example: 0.98
            factors:
              type: object
              additionalProperties:
                type: number
        data_freshness:
          type: string
          format: date-time
          description: ISO timestamp of the data's as-of datetime.
          example: "2026-02-21T10:30:00Z"
        billing_code:
          type: string
          description: Internal billing classification for this request type.
          example: ticker_returns_v2
        cache_status:
          type: string
          enum: [HIT, MISS, BYPASS]
          description: Whether this response was served from cache.
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Machine-readable error code.
          example: TICKER_NOT_FOUND
        message:
          type: string
          description: Human-readable error description with suggested action.
          example: "Ticker 'XYZABC' not found in universe 'uni_mc_3000'. Check ticker spelling or try a different universe."
        code:
          type: integer
          description: HTTP status code.
          example: 404
        details:
          type: object
          properties:
            field:
              type: string
              example: ticker
            received:
              type: string
              example: XYZABC
            universe:
              type: string
              example: uni_mc_3000
            suggestion:
              type: string
              example: "Check for typos or use ticker search endpoint"
...
    InsufficientBalance:
      type: object
      description: Returned when the account prepaid balance is exhausted (HTTP 402).
      properties:
        error:
          type: string
          enum: [INSUFFICIENT_BALANCE]
          example: INSUFFICIENT_BALANCE
        message:
          type: string
          example: "Insufficient balance. Current balance: $0.00. Top up at https://riskmodels.app/billing"
        balance_usd:
          type: number
          format: float
          example: 0.0
        top_up_url:
          type: string
          format: uri
          example: https://riskmodels.app/billing
    RateLimitExceeded:
      type: object
      description: Returned when the per-minute rate limit is exceeded (HTTP 429).
      properties:
        error:
          type: string
          enum: [RATE_LIMIT_EXCEEDED]
          example: RATE_LIMIT_EXCEEDED
        message:
          type: string
          example: "Rate limit exceeded. Retry after 2026-03-08T12:34:56Z"
        retry_after:
          type: string
          format: date-time
          description: ISO timestamp after which the client may retry.
          example: "2026-03-08T12:34:56Z"
        limit:
          type: integer
          description: Requests allowed per minute for this key.
          example: 60
security:
  - BearerAuth: []
paths:
...
  /rankings/{ticker}:
    get:
      summary: Cross-sectional rankings snapshot
      description: >
        Latest V3 cross-sectional ranks from `security_history` for the ticker’s symbol.
        Optional query filters narrow to one metric, cohort, and/or window; omit them for the full grid.
        See `#/components/schemas/RankingMetricKeys` for key patterns. `rank_percentile` is 100 = best.
      operationId: getRankingsByTicker
      tags: [Risk Metrics]
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
          example: AAPL
        - name: metric
          in: query
          required: false
          schema:
            type: string
            enum:
              - mkt_cap
              - gross_return
              - sector_residual
              - subsector_residual
              - er_l1
              - er_l2
              - er_l3
        - name: cohort
          in: query
          required: false
          schema:
            type: string
            enum: [universe, sector, subsector]
        - name: window
          in: query
          required: false
          schema:
            type: string
            enum: [1d, 21d, 63d, 252d]
      responses:
        "200":
          description: Rankings and lineage metadata.
          headers:
            X-API-Cost-USD:
              schema: { type: string }
            X-Data-Fetch-Latency-Ms:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticker: { type: string }
                  symbol: { type: string }
                  teo:
                    type: string
                    format: date
                    nullable: true
                  rankings:
                    type: array
                    items:
                      type: object
                      properties:
                        metric: { type: string }
                        cohort: { type: string }
                        window: { type: string }
                        rank_ordinal:
                          type: integer
                          nullable: true
                        cohort_size:
                          type: integer
                          nullable: true
                        rank_percentile:
                          type: number
                          nullable: true
                  filters:
                    type: object
                    properties:
                      metric: { type: string, nullable: true }
                      cohort: { type: string, nullable: true }
                      window: { type: string, nullable: true }
                  _metadata:
                    type: object
                    additionalProperties: true
                  _agent:
                    $ref: "#/components/schemas/AgentMeta"
        "400":
          description: Invalid query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid Bearer token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "402":
          description: Insufficient balance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InsufficientBalance"
        "404":
          description: Ticker not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitExceeded"
...
  /tickers:
    get:
      summary: Ticker universe search
      description: >
        List tickers in the universe or search by name/symbol. Free endpoint (no charge).
        Use the `mag7` flag to retrieve the MAG7 tickers. Use `include_metadata` for
        sector and ETF assignment per ticker.
      operationId: getTickers
      tags: [Utility]
      security: []
      parameters:
        - name: array
          in: query
          required: false
          schema:
            type: string
            enum: [ticker, teo]
          description: "`ticker` returns all ticker symbols. `teo` returns valid trading dates."
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search string matched against ticker symbol or company name.
          example: NVDA
        - name: mag7
          in: query
          required: false
          schema:
            type: boolean
          description: Return only MAG7 tickers (AAPL, MSFT, NVDA, GOOGL, AMZN, META, TSLA).
        - name: include_metadata
          in: query
          required: false
          schema:
            type: boolean
          description: Include sector, sector ETF, and subsector ETF per ticker.
      responses:
        "200":
          description: Ticker list or search results.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tickers:
                    type: array
                    items:
                      type: string
                    example: ["AAPL", "MSFT", "NVDA"]
                  metadata:
                    type: object
                    nullable: true
                    additionalProperties:
                      type: object
                      properties:
                        name:
                          type: string
                        sector:
                          type: string
                        sector_etf:
                          type: string
                        subsector_etf:
                          type: string
...
  /user/billing-config:
    get:
      summary: Auto-refill billing preferences
      description: >
        Current auto-refill flags and amounts from `agent_accounts`. Same semantics as portal settings;
        use `PATCH /balance` for the alternate auto-refill field names (`enabled`, `refill_amount_usd`).
      operationId: getUserBillingConfig
      tags: [Account]
      responses:
        "200":
          description: Billing preferences.
          content:
            application/json:
              schema:
                type: object
                properties:
                  auto_top_up: { type: boolean }
                  auto_top_up_amount: { type: number }
                  auto_top_up_threshold: { type: number }
                  has_payment_method: { type: boolean }
                  allowed_refill_amounts:
                    type: array
                    items: { type: number }
                  threshold_bounds:
                    type: object
                    properties:
                      min: { type: number }
                      max: { type: number }
        "401":
          description: Unauthorized.
        "404":
          description: Agent account not found.
        "500":
          description: Server error.
    patch:
      summary: Update auto-refill billing preferences
      description: >
        Partial update of `auto_top_up`, `auto_top_up_amount` (must be 20, 50, or 100), and/or
        `auto_top_up_threshold` (USD, 5–50). Requires a saved payment method to enable auto-refill.
      operationId: patchUserBillingConfig
      tags: [Account]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                auto_top_up: { type: boolean }
                auto_top_up_amount:
                  type: number
                  enum: [20, 50, 100]
                auto_top_up_threshold:
                  type: number
                  minimum: 5
                  maximum: 50
      responses:
        "200":
          description: Updated fields.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  auto_top_up: { type: boolean }
                  auto_top_up_amount: { type: number }
                  auto_top_up_threshold: { type: number }
        "400":
          description: Validation or missing payment method.
        "401":
          description: Unauthorized.
        "500":
          description: Server error.
...